requirements.txt
django==3.0.3
django-heroku==0.3.1
gunicorn==20.0.4
psycopg2-binary==2.8.4
requests==2.22.0
django-crispy-forms==1.8.1
importlib-metadata==1.4.0
python-dotenv==0.10.5